Move ConvertSpaceNames to network package
