--- worker/discoverspaces/export_test.go
@@ -1,6 +1,0 @@
-// Copyright 2016 Canonical Ltd.
-// Licensed under the AGPLv3, see LICENCE file for details.
-
-package discoverspaces
-
-var ConvertSpaceName = convertSpaceName
